[charts-pro] Split `ChartAxisZoomSlider` into smaller files
